Add Leaderboard Page

+++ src/server/GliderView.Data/Models/Flight.cs
@@ -30,7 +30,10 @@ namespace GliderView.Data.Models
 
         // Statistics
         public int? MaxAltitude { get; set; }
-        public int? DistanceTraveled { get; set; }
+        /// <summary>
+        /// Kilometers
+        /// </summary>
+        public float? DistanceTraveled { get; set; }
         public int? AltitudeGained { get; set; }
         public int? ReleaseHeight { get; set; }
         public int? PatternEntryAltitude { get; set; }
